fix handler, not use state
