+++ src/components/mylist/mylist.jsx
@@ -34,7 +34,7 @@ const MyList = ({films}) => {
           </a>
         </div>
         <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
+          <p>© 2020 What to watch Ltd.</p>
         </div>
       </footer>
     </div>
